resume_job_matcher/requirements.txt
PyPDF2
python-docx
nltk
scikit-learn
streamlit
